Issue with module imports temporarily resolved
There seems to be an issue with importing the new classes I added to host. After lots of digging and testing this was the only thing I could come up with that would get it to function. This is obviously a temporary fix, will see what I can do to get memory, hardware, and mmu into their own files.